Add next ID tool
- It prints the next show and venue IDs for human editing of the raw files.
- This also removes the tools added with #478 and #481
